+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added mfc140 and fonts installing localization
+
+### Fixed
+
+- Fixed check button style for newly made sessions
+- Fixed repairer's NaN progress
+- Fixed game session selection when current one is removed
+- Fixed mfc140 and fonts installing
+
+### Changed
+
+- Updated Spanish
+- Updated Hungarian
+
 ## [1.0.0] - 11.06.2023
 
 🚀 Initial release
@@ -14,3 +30,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <br>
 
 [unreleased]: https://github.com/an-anime-team/an-anime-borb-launcher/compare/1.0.0...next
+[1.0.0]: https://github.com/an-anime-team/an-anime-borb-launcher/releases/tag/1.0.0